Move get_column() to ColumnId
