+++ libs/exo-sql/src/lib.rs
@@ -63,13 +63,13 @@ pub use asql::{
 pub use sql::{
     array_util::{self, ArrayEntry},
     column::Column,
-    database::ColumnId,
     database::Database,
     database::TableId,
     database_client::{DatabaseClient, LOCAL_CONNECTION_POOL_SIZE, LOCAL_URL},
     limit::Limit,
     offset::Offset,
     order::Ordering,
+    physical_column::ColumnId,
     physical_column::{FloatBits, IntBits, PhysicalColumn, PhysicalColumnType},
     physical_table::PhysicalTable,
     predicate::{CaseSensitivity, ParamEquality, Predicate},
