+++ libs/exo-sql/src/sql/physical_column.rs
@@ -228,3 +228,37 @@ impl PhysicalColumnType {
         }
     }
 }
+
+#[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Eq, Copy, Hash)]
+pub struct ColumnId {
+    pub table_id: TableId,
+    pub column_index: usize,
+}
+
+impl ColumnId {
+    pub fn new(table_id: TableId, column_index: usize) -> ColumnId {
+        ColumnId {
+            table_id,
+            column_index,
+        }
+    }
+
+    pub fn get_column<'a>(&self, database: &'a Database) -> &'a PhysicalColumn {
+        &database.get_table(self.table_id).columns[self.column_index]
+    }
+}
+
+impl PartialOrd for ColumnId {
+    fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
+        Some(self.cmp(other))
+    }
+}
+
+impl Ord for ColumnId {
+    fn cmp(&self, other: &Self) -> std::cmp::Ordering {
+        fn tupled(a: &ColumnId) -> (usize, usize) {
+            (a.table_id.arr_idx(), a.column_index)
+        }
+        tupled(self).cmp(&tupled(other))
+    }
+}
